# contracts/marketplace-customer.openapi.yaml
openapi: 3.1.0
info:
  title: Tellbook Marketplace Customer Data API
  version: 1.0.0
servers:
  - url: /v1/marketplace
security:
  - marketplaceSession: []
paths:
...
  /bookings/claims:
    post:
      operationId: claimMarketplaceBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [booking_token]
              properties:
                booking_token: { type: string, minLength: 1 }
      responses:
        '200':
          description: Booking atomically linked to the matching verified customer.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookingClaimResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Verified customer identity does not match the booking contact. }
        '409': { description: Booking is already owned by another customer. }
...
components:
  securitySchemes:
    marketplaceSession:
      type: apiKey
      in: cookie
      name: tellbook_marketplace_session
...
  responses:
...
    Unauthorized:
      description: Marketplace customer session is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    BookingClaimResponse:
      type: object
      additionalProperties: false
      required: [booking_id]
      properties:
        booking_id: { type: string, format: uuid }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
...